--- Mock CCC/3 Mock CCC 18 Contest 2 J1 An Arithmetic Problem/src/Main.java
@@ -1,13 +1,0 @@
-import java.io.*;
-public class Main {
-	/* 
-	 * Copy-pasting code is NOT cool! Please do not copy and paste my code as a submission to DMOJ.
-	 * github.com/jdabtieu/competitive-programming
-	 */
-	public static void main(String[] args) throws IOException {
-		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
-		String[] a = br.readLine().split(" ");
-		System.out.println(Integer.parseInt(a[0])+Integer.parseInt(a[2])==Integer.parseInt(a[4]) ? "True" : "False");
-	}
-
-}
