--- Mock CCC/3 Mock CCC 18 Contest 3 J1 A Sum Problem/src/Main.java
@@ -1,17 +1,0 @@
-import java.util.*;
-import java.io.*;
-public class Main {
-	/* 
-	 * Copy-pasting code is NOT cool! Please do not copy and paste my code as a submission to DMOJ.
-	 * github.com/jdabtieu/competitive-programming
-	 */
-	public static void main(String[] args) throws IOException {
-	    	BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
-	    	int sum = 0;
-		for (int i = Integer.parseInt(br.readLine()); i > 0; i--) {
-			sum += Integer.parseInt(br.readLine());
-		}
-		System.out.println(sum);
-
-	}
-}
